+++ CHANGELOG.md
@@ -14,6 +14,8 @@ CHANGELOG for FlatCAM beta
 - fixed the Tcl COmmand Panelize, Excellon panelization section
 - Fixed an issue in Tool Calibration export_excellon method call
 - PEP8 corrections all over the app
+- made sure that the object selection will not work while in Editors or in the App Tools
+- some minor changes to strings and icons
 
 1.11.2020